Feat: add orders service (fix update order later).

+++ src/Services/Basket/Basket.Api/Models/Cart.cs
@@ -4,6 +4,6 @@
     {
         public string UserName { get; set; } = default!;
         public List<CartItem> CartItems { get; set; } = new();
-        public decimal TotalPrice { get => CartItems.Sum(x => x.Price * x.Quantity); }
+        public double TotalPrice { get => CartItems.Sum(x => x.Price * x.Quantity); }
     }
 }
